Add experimental GpsModel as a layer on top of OpenGps

+++ lib/src/main/java/net/dhleong/opengps/OpenGps.java
@@ -59,7 +59,7 @@ public class OpenGps {
             });
     }
 
-    public Observable<AeroObject> anyNear(double lat, double lng, double range) {
+    public Observable<AeroObject> anyNear(double lat, double lng, float range) {
         // TODO
         return Observable.empty();
     }
